Add USB bluetooth config to install.sh and lib.sh
Updated script to include automatic Bluetooth USB device detection and attachment to core-USB. (Thought I'd give a brief assist as this was in the post-1.0 release TODO.)

Also: added a new function filter_bluetooth_devices in lib.sh for filtering out Bluetooth devices from the system USB list. This function is used in the main script to attach the filtered BT devices.